+++ android/app/src/main/java/org/ebookdroid/droids/RtfContext.java
@@ -36,7 +36,6 @@ public class RtfContext extends PdfContext {
             }
         }
 
-        MuPdfDocument muPdfDocument = new MuPdfDocument(this, MuPdfDocument.FORMAT_PDF, cacheFile.getPath(), password);
-        return muPdfDocument;
+        return openTextDoc(fileName, cacheFile.getPath(), password);
     }
 }